Fix CI failures present since 2026-09-04
- mypy: annotations / float() casts in test_analysis_readiness.py,
  reference_mixture.py and fst.py (clean on Python 3.11 and 3.12).
- Parity job: dsuite has never been a conda package, so the environment
  could not solve. Build Dsuite from a pinned upstream commit instead.
- Dsuite parity test never ran: fix the lower-case jackal outgroup, use
  site-count blocks on the one-contig simulation, and require every
  Dtrios trio to match (three ingroups give one trio, not three).
- Simulated cohorts: newer tskit refuses a VCF site at position 0; write
  it at 1.
- Re-bless population_genomics.json and introgression.json. They were
  blessed on a machine whose simulator produced a different cohort and
  never ran in CI; the 2026-08-17 code and current code give identical
  results on the same simulated input.
- tests/fixtures/real_bridge/bridge.vcf.gz was never committed (hidden by
  the *.vcf.gz ignore rule): skip its tests with that reason and exempt
  fixture VCFs from the rule.

+++ src/canidae/analysis/reference_mixture.py
@@ -231,8 +231,8 @@ def infer_dog_fraction(
         if observed is None or abs(difference) < min_difference:
             continue
         called += 1
-        numerator += (dosage(observed) - coyote) * difference
-        denominator += difference * difference
+        numerator += float((dosage(observed) - coyote) * difference)
+        denominator += float(difference * difference)
     if called < min_called or denominator == 0:
         return None, called, None
     estimate = max(0.0, min(1.0, numerator / denominator))
